requirements.txt
pint
numpy
scipy
streamlit 
pyyaml 
pandas